refactor: 지연로딩 적용 및 properties 파일 수정
