+++ backend/sin/src/main/java/sin/sin/domain/productDescription/ProductDescription.java
@@ -2,6 +2,7 @@ package sin.sin.domain.productDescription;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -23,7 +24,7 @@ public class ProductDescription {
     @Column(name="product_description_id", nullable = false)
     private Long id;
 
-    @OneToOne
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(nullable = false, name = "product_id")
     private Product product;
