+++ backend/sin/src/main/java/sin/sin/domain/productReviewImg/ProductReviewImg.java
@@ -1,4 +1,4 @@
-package sin.sin.domain.roductReviewImg;
+package sin.sin.domain.productReviewImg;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
