+++ backend/sin/src/main/java/sin/sin/dto/NotificationResponse.java
@@ -3,8 +3,6 @@ package sin.sin.dto;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.sql.Timestamp;
-
 @Getter
 @Builder
 public class NotificationResponse {
