+++ backend/sin/src/main/resources/application.properties
@@ -1,15 +1,13 @@
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
-spring.datasource.url=jdbc:h2:tcp://localhost/~/testdb
+spring.datasource.url=jdbc:h2:mem:testdb;
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
+#update로는 import.sql에 있는 insert가 효과 없는듯?
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.format_sql=true
-
-#spring.sql.init.schema-locations=classpath:notification.sql
-#spring.sql.init.mode=always
+spring.jpa.show-sql=true
