--- backend/sin/src/main/resources/notification.sql
@@ -1,41 +1,0 @@
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목1','내용1','글쓴이1');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목2','내용2','글쓴이2');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목3','내용3','글쓴이3');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목4','내용4','글쓴이4');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목5','내용5','글쓴이5');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목6','내용6','글쓴이6');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목7','내용7','글쓴이7');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목8','내용8','글쓴이8');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목9','내용9','글쓴이9');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목10','내용10','글쓴이10');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목11','내용11','글쓴이11');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목12','내용12','글쓴이12');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목13','내용13','글쓴이13');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목14','내용14','글쓴이14');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목15','내용15','글쓴이15');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목16','내용16','글쓴이16');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목17','내용17','글쓴이17');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목18','내용18','글쓴이18');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목19','내용19','글쓴이19');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목20','내용20','글쓴이20');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목21','내용21','글쓴이21');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목22','내용22','글쓴이22');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목23','내용23','글쓴이23');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목24','내용24','글쓴이24');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목25','내용25','글쓴이25');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목26','내용26','글쓴이26');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목27','내용27','글쓴이27');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목28','내용28','글쓴이28');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목29','내용29','글쓴이29');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목30','내용30','글쓴이30');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목31','내용31','글쓴이31');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목32','내용32','글쓴이32');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목33','내용33','글쓴이33');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목34','내용34','글쓴이34');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목35','내용35','글쓴이35');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목36','내용36','글쓴이36');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목37','내용37','글쓴이37');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목38','내용38','글쓴이38');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목39','내용39','글쓴이39');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목40','내용40','글쓴이40');
-INSERT INTO Notification (TITLE, CONTENT, WRITER) VALUES('제목41','내용41','글쓴이41');
